Organize scripts and events & fix error
cloudAI v1.0.5-dev

-> From last PR Merged to dev
--> Add Ollama Support

-> Fix error that i've missed from PR
-> Organize Events, Scripts
-> All files now in src folder
