+++ package.json
@@ -1,9 +1,9 @@
 {
   "name": "cloudai",
-  "version": "1.0.0",
+  "version": "1.0.5",
   "main": "index.js",
   "scripts": {
-    "start": "node ."
+    "start": "node src/index.js"
   },
   "author": "gunawan092w",
   "license": "MIT",
@@ -12,6 +12,6 @@
     "@google/genai": "^1.5.1",
     "discord.js": "^14.20.0",
     "ollama": "^0.5.16",
-    "openai": "^5.6.0",
+    "openai": "^5.6.0"
   }
 }
