+++ src/config.js
@@ -1,0 +1,3 @@
+const { bot, ai } = require("../config.json");
+module.exports = { bot, ai };
+
