+++ src/events/ready.js
@@ -1,0 +1,10 @@
+const { Events } = require('discord.js');
+
+module.exports = {
+	name: Events.ClientReady,
+	once: true,
+	execute(client) {
+		console.log(`Logged in as ${client.user.tag}`);
+	},
+};
+
