+++ src/scripts/chatgptAPI.js
@@ -1,0 +1,13 @@
+const { OpenAI } = require('openai');
+const { ai } = require('../config');
+
+const chatGPT = new OpenAI({apiKey: ai.chatgpt.apiKey});
+
+async function chatgptResponse(prompt) {
+    const response = await chatgptAI.responses.create({
+        model: "o4-mini",
+        instructions: "You are a discord AI bot. Do not exceed 4096 characters.",
+        input: prompt
+    }); return response.output_text;
+}; module.exports = { chatgptResponse };
+
